Rename core/auth/login folder to login-page.
