+++ src/app/core/auth/auth.module.ts
@@ -2,7 +2,7 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { HttpClientModule } from '@angular/common/http';
 import { ReactiveFormsModule } from '@angular/forms';
-import { LoginPageComponent } from './pages/login/login-page.component';
+import { LoginPageComponent } from './pages/login-page/login-page.component';
 import { RouterModule } from '@angular/router';
 import { LogoutComponent } from './pages/logout/logout.component';
 import { RegisterPageComponent } from './pages/register-page/register-page.component';
